jetson-nano-source-code/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.12
PyYAML>=5.4.1
tqdm>=4.62.3
tensorboard>=2.8.0
pillow>=8.3.1
numpy>=1.21.2
